chore: Side, Day 컴포넌트 Table 폴더로 이동, 폴더 별 css 관리
