+++ frontend/src/components/Calendar/Table/CalendarDay.tsx
@@ -1,6 +1,6 @@
-import { dayStyle } from '../index.css';
 import type { Time } from '../types';
 import { CalendarCell } from './CalendarCell';
+import { dayStyle } from './index.css';
 
 export const CalendarDay = ({ holiday = false }: { holiday?: boolean }) => {
   const TIMES: Time[] = new Array(24).fill(0).map((_, i) => i);
